+++ cypress/integration/edit-ontology.spec.js
@@ -11,11 +11,13 @@ import getStyling from '../fixtures/getStyling'
 import linkAutocomplete from '../fixtures/linkAutocomplete'
 import linkSearch from '../fixtures/linkSearch'
 import { ROUTE_EDIT_ONTOLOGY } from '../../src/constants/routes'
+import showTourLs from '../fixtures/showTourLs'
 
 context('Edit ontology', () => {
   beforeEach(() => {
     cy.visit('/')
     cy.get('#accept-all-btn').click()
+    window.localStorage.setItem('showTour', showTourLs)
   })
 
   describe('Edit ontology', () => {
@@ -88,7 +90,7 @@ context('Edit ontology', () => {
       cy.get('#email').type('valid@email.com')
       cy.get('#password').type('password')
 
-      cy.get('.auth-button').click()
+      cy.get('#auth-login-button').click()
 
       cy.wait('@postLogin')