+++ cypress/integration/network-graph-options.spec.js
@@ -6,11 +6,13 @@ import getStyling from '../fixtures/getStyling'
 import linkAutocomplete from '../fixtures/linkAutocomplete'
 import linkSearch from '../fixtures/linkSearch'
 import { ROUTE_NETWORK_GRAPH_OPTIONS } from '../../src/constants/routes'
+import showTourLs from '../fixtures/showTourLs'
 
 context('Network graph options', () => {
   beforeEach(() => {
     cy.visit('/')
     cy.get('#accept-all-btn').click()
+    window.localStorage.setItem('showTour', showTourLs)
   })
 
   describe('Network graph options', () => {
@@ -58,7 +60,7 @@ context('Network graph options', () => {
       cy.get('#email').type('valid@email.com')
       cy.get('#password').type('password')
 
-      cy.get('.auth-button').click()
+      cy.get('#auth-login-button').click()
 
       cy.wait('@postLogin')