+++ cypress/integration/network-styling.spec.js
@@ -9,11 +9,13 @@ import saveStyling from '../fixtures/saveStyling'
 import linkAutocomplete from '../fixtures/linkAutocomplete'
 import linkSearch from '../fixtures/linkSearch'
 import { ROUTE_STYLING } from '../../src/constants/routes'
+import showTourLs from '../fixtures/showTourLs'
 
 context('Network styling', () => {
   beforeEach(() => {
     cy.visit('/')
     cy.get('#accept-all-btn').click()
+    window.localStorage.setItem('showTour', showTourLs)
   })
 
   describe('Network styling nodes', () => {