+++ src/__test__/components/Sidebar.test.js
@@ -1,27 +1,23 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 import toJson from 'enzyme-to-json'
-import { useRouter } from 'next/router'
 import Sidebar from '../../components/Sidebar'
-import {
-  ROUTE_CUSTOM_QUERY,
-  ROUTE_ELEMENTS_FILTER,
-  ROUTE_ELEMENTS_SELECTION,
-  ROUTE_EDIT_ONTOLOGY,
-  ROUTE_EXPORT,
-  ROUTE_NETWORK_GRAPHS,
-  ROUTE_NETWORK_GRAPH_OPTIONS,
-  ROUTE_NODE_NEIGHBOURHOOD,
-  ROUTE_SEARCH, ROUTE_SETTINGS,
-  ROUTE_SHORTEST_PATH,
-  ROUTE_STYLING,
-  ROUTE_SYNONYMS
-} from '../../constants/routes'
 
-jest.mock('next/router')
+import {
+  SIDEBAR_VIEW_CUSTOM_QUERY,
+  SIDEBAR_VIEW_EDIT_ONTOLOGY,
+  SIDEBAR_VIEW_ELEMENTS_FILTER,
+  SIDEBAR_VIEW_ELEMENTS_SELECTION,
+  SIDEBAR_VIEW_ENTRY_SEARCH, SIDEBAR_VIEW_EXPORT,
+  SIDEBAR_VIEW_GRAPHS, SIDEBAR_VIEW_GRAPH_OPTIONS,
+  SIDEBAR_VIEW_NEIGHBOURHOOD, SIDEBAR_VIEW_SHORTEST_PATH,
+  SIDEBAR_VIEW_STYLING, SIDEBAR_VIEW_SYNONYMS
+} from '../../constants/views'
 
 const setup = ({
   isSidebarOpen,
+  sidebarView,
+  user
 }) => {
   const props = {
     isSidebarOpen,
@@ -35,7 +31,12 @@ const setup = ({
         label: 'Search',
       }
     },
+    sidebarView,
     updateStoreValue: jest.fn(),
+  }
+
+  if (user) {
+    props.user = user
   }
 
   const component = shallow(<Sidebar {...props} />)
@@ -51,189 +52,152 @@ describe('Sidebar', () => {
     const {
       component
     } = setup({
-      isSidebarOpen: false
+      isSidebarOpen: false,
+      sidebarView: SIDEBAR_VIEW_ENTRY_SEARCH,
+    })
+
+    expect(toJson(component)).toMatchSnapshot()
+  })
+
+  it('should match snapshot when sidebar open and SIDEBAR_VIEW_ENTRY_SEARCH and isGuest', () => {
+    const {
+      component
+    } = setup({
+      isSidebarOpen: false,
+      sidebarView: SIDEBAR_VIEW_ENTRY_SEARCH,
+      user: { isGuest: true }
     })
 
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it('should match snapshot when sidebar open and SIDEBAR_VIEW_ENTRY_SEARCH', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_SEARCH
-    }))
-
     const {
       component
     } = setup({
       isSidebarOpen: true,
+      sidebarView: SIDEBAR_VIEW_ENTRY_SEARCH,
     })
 
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it('should match snapshot when sidebar open and SIDEBAR_VIEW_GRAPHS', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_NETWORK_GRAPHS
-    }))
-
     const {
       component
     } = setup({
       isSidebarOpen: true,
+      sidebarView: SIDEBAR_VIEW_GRAPHS,
     })
 
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it('should match snapshot when sidebar open and SIDEBAR_VIEW_GRAPH_OPTIONS', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_NETWORK_GRAPH_OPTIONS
-    }))
-
     const {
       component
     } = setup({
       isSidebarOpen: true,
+      sidebarView: SIDEBAR_VIEW_GRAPH_OPTIONS,
     })
 
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it('should match snapshot when sidebar open and SIDEBAR_VIEW_NEIGHBOURHOOD', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_NODE_NEIGHBOURHOOD
-    }))
-
     const {
       component
     } = setup({
       isSidebarOpen: true,
+      sidebarView: SIDEBAR_VIEW_NEIGHBOURHOOD,
     })
 
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it('should match snapshot when sidebar open and SIDEBAR_VIEW_ELEMENTS_SELECTION', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_ELEMENTS_SELECTION
-    }))
-
     const {
       component
     } = setup({
       isSidebarOpen: true,
+      sidebarView: SIDEBAR_VIEW_ELEMENTS_SELECTION,
     })
 
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it('should match snapshot when sidebar open and SIDEBAR_VIEW_SHORTEST_PATH', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_SHORTEST_PATH
-    }))
-
     const {
       component
     } = setup({
       isSidebarOpen: true,
-    })
-
-    expect(toJson(component)).toMatchSnapshot()
-  })
-
-  it('should match snapshot when sidebar open and SIDEBAR_VIEW_SETTINGS', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_SETTINGS
-    }))
-
-    const {
-      component
-    } = setup({
-      isSidebarOpen: true,
+      sidebarView: SIDEBAR_VIEW_SHORTEST_PATH,
     })
 
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it('should match snapshot when sidebar open and ROUTE_ELEMENTS_FILTER', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_ELEMENTS_FILTER
-    }))
-
     const {
       component
     } = setup({
       isSidebarOpen: true,
+      sidebarView: SIDEBAR_VIEW_ELEMENTS_FILTER,
     })
 
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it('should match snapshot when sidebar open and SIDEBAR_VIEW_CUSTOM_QUERY', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_CUSTOM_QUERY
-    }))
-
     const {
       component
     } = setup({
       isSidebarOpen: true,
+      sidebarView: SIDEBAR_VIEW_CUSTOM_QUERY,
     })
 
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it('should match snapshot when sidebar open and SIDEBAR_VIEW_EXPORT', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_EXPORT
-    }))
-
     const {
       component
     } = setup({
       isSidebarOpen: true,
+      sidebarView: SIDEBAR_VIEW_EXPORT,
     })
 
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it('should match snapshot when sidebar open and SIDEBAR_VIEW_STYLING', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_STYLING
-    }))
-
     const {
       component
     } = setup({
       isSidebarOpen: true,
+      sidebarView: SIDEBAR_VIEW_STYLING,
     })
 
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it('should match snapshot when sidebar open and SIDEBAR_VIEW_EDIT_ONTOLOGY', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_EDIT_ONTOLOGY
-    }))
-
     const {
       component
     } = setup({
       isSidebarOpen: true,
+      sidebarView: SIDEBAR_VIEW_EDIT_ONTOLOGY,
     })
 
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it('should match snapshot when sidebar open and SIDEBAR_VIEW_SYNONYMS', () => {
-    useRouter.mockImplementation(() => ({
-      view: ROUTE_SYNONYMS
-    }))
-
     const {
       component
     } = setup({
       isSidebarOpen: true,
+      sidebarView: SIDEBAR_VIEW_SYNONYMS,
     })
 
     expect(toJson(component)).toMatchSnapshot()