+++ src/components/NetworkGraphList.js
@@ -2,21 +2,63 @@ import { connect } from 'redux-zero/react'
 import PropTypes from 'prop-types'
 import { useTranslation } from 'react-i18next'
 import { Button } from 'primereact/button'
+import Joyride from 'react-joyride'
 import actions from '../store/actions'
-import { SIDEBAR_VIEW_GRAPHS } from '../constants/views'
-import { OPERATION_TYPE_DELETE, OPERATION_TYPE_UPDATE } from '../constants/store'
+import { NETWORK_VIEW_DATATABLE, SIDEBAR_VIEW_GRAPHS } from '../constants/views'
+import { OPERATION_TYPE_DELETE, OPERATION_TYPE_OBJECT_ADD, OPERATION_TYPE_UPDATE } from '../constants/store'
 
 const NetworkGraphList = ({
   updateStoreValue,
   currentGraph,
-  graphData
+  graphData,
+  showTour
 }) => {
   const { t } = useTranslation()
 
   const graphViewsKeys = Object.keys(graphData)
 
+  const steps = [
+    {
+      target: '.vis-up',
+      content: t('introNavigateButtons'),
+      placement: 'top',
+      disableBeacon: true
+    },
+    {
+      target: '.vis-zoomExtends',
+      content: t('introNavigateZoom'),
+      placement: 'top',
+      disableBeacon: true
+    },
+    {
+      target: '#navbar-datatable-btn',
+      content: t('introNavigateDatatable'),
+      placement: 'bottom',
+      disableBeacon: true
+    }
+  ]
+
+  const handleJoyrideCallback = (data) => {
+    const { status } = data
+
+    if (status === 'finished') {
+      localStorage.setItem('showTour', JSON.stringify({ ...showTour, navigate: 'false' }))
+      updateStoreValue(['showTour'], OPERATION_TYPE_OBJECT_ADD, { navigate: 'false' })
+      updateStoreValue(['networkVisualisation'], OPERATION_TYPE_UPDATE, NETWORK_VIEW_DATATABLE)
+    }
+  }
+
   return (
     <>
+      {showTour.navigate !== 'false' && (
+        <Joyride
+          callback={handleJoyrideCallback}
+          steps={steps}
+          disableScrolling
+          hideBackButton
+          locale={{ close: t('next') }}
+        />
+      )}
       <h1 className="sidebar-main-title">
         {t(SIDEBAR_VIEW_GRAPHS)}
       </h1>
@@ -73,14 +115,17 @@ NetworkGraphList.propTypes = {
   updateStoreValue: PropTypes.func.isRequired,
   graphData: PropTypes.shape().isRequired,
   currentGraph: PropTypes.string.isRequired,
+  showTour: PropTypes.shape().isRequired,
 }
 
 const mapToProps = ({
   graphData,
-  currentGraph
+  currentGraph,
+  showTour
 }) => ({
   graphData,
-  currentGraph
+  currentGraph,
+  showTour
 })
 
 export default connect(