+++ src/components/NotesList.js
@@ -12,6 +12,7 @@ import { Button } from 'primereact/button'
 import { Calendar } from 'primereact/calendar'
 import { Accordion, AccordionTab } from 'primereact/accordion'
 import { MultiSelect } from 'primereact/multiselect'
+import Joyride from 'react-joyride'
 import { SIDEBAR_VIEW_NOTES } from '../constants/views'
 import actions from '../store/actions'
 import NotesListNote from './NotesListNote'
@@ -20,7 +21,7 @@ import addNodesBorders from '../utils/networkStyling/addNodesBorders'
 import { SORT_FIELDS, MIN_DATE } from '../constants/notes'
 import getEdgeIds from '../utils/nodesEdgesUtils/getEdgeIds'
 import getEdge from '../utils/nodesEdgesUtils/getEdge'
-import { OPERATION_TYPE_UPDATE } from '../constants/store'
+import { OPERATION_TYPE_OBJECT_ADD, OPERATION_TYPE_UPDATE } from '../constants/store'
 import updateHighlightedElement from '../utils/networkStyling/updateHighlightedElement'
 
 const NotesList = ({
@@ -31,7 +32,8 @@ const NotesList = ({
   selectedNotesType,
   noteElementId,
   classesFromApi,
-  nodesDropdownLabels
+  nodesDropdownLabels,
+  showTour
 }) => {
   const { t } = useTranslation()
 
@@ -122,8 +124,58 @@ const NotesList = ({
     })
   }).filter((edge) => edge.label)
 
+  const steps = [
+    {
+      target: '#notes-select',
+      content: t('introNotesType'),
+      placement: 'top',
+      disableBeacon: true
+    },
+    {
+      target: '#notes-select-element',
+      content: t('introNotesElement'),
+      placement: 'top',
+      disableBeacon: true
+    }
+  ]
+
+  const handleJoyrideCallback = (data) => {
+    const { status, index } = data
+
+    if (index === 1) {
+      updateStoreValue(['isNodeSelectable'], OPERATION_TYPE_UPDATE, true)
+      updateStoreValue(['isEdgeSelectable'], OPERATION_TYPE_UPDATE, false)
+      updateStoreValue(['noteElementId'], OPERATION_TYPE_UPDATE, undefined)
+      updateStoreValue(['selectedNotesType'], OPERATION_TYPE_UPDATE, 'node')
+    }
+
+    if (index === 2) {
+      updateHighlightedElement({
+        updateStoreValue,
+        id: '2',
+        type: selectedNotesType
+      })
+    }
+
+    if (status === 'finished') {
+      localStorage.setItem('showTour', JSON.stringify({ ...showTour, notes: 'false' }))
+      updateStoreValue(['showTour'], OPERATION_TYPE_OBJECT_ADD, { notes: 'false' })
+      document.getElementById('sidebar-button-synonyms').click()
+    }
+  }
+
   return (
     <>
+      {showTour.notes !== 'false' && (
+      <Joyride
+        callback={handleJoyrideCallback}
+        steps={steps}
+        disableScrolling
+        hideBackButton
+        locale={{ close: t('next') }}
+      />
+      )}
+
       <h1 className="sidebar-main-title">
         {t(SIDEBAR_VIEW_NOTES)}
       </h1>
@@ -338,6 +390,7 @@ NotesList.propTypes = {
   updateStoreValue: PropTypes.func.isRequired,
   classesFromApi: PropTypes.shape().isRequired,
   nodesDropdownLabels: PropTypes.arrayOf(PropTypes.shape).isRequired,
+  showTour: PropTypes.shape().isRequired,
 }
 
 NotesList.defaultProps = {
@@ -351,7 +404,8 @@ const mapToProps = ({
   edgesNotes,
   noteElementId,
   classesFromApi,
-  nodesDropdownLabels
+  nodesDropdownLabels,
+  showTour
 }) => ({
   notes,
   selectedNotesType,
@@ -359,7 +413,8 @@ const mapToProps = ({
   edgesNotes,
   noteElementId,
   classesFromApi,
-  nodesDropdownLabels
+  nodesDropdownLabels,
+  showTour
 })
 
 export default connect(